==> Instacart Basket Analysis/03 Scripts/01 Data Wrangling & Subsetting.ipynb <==
# 01 Data Wrangling & Subsetting

    01 Import data and libraries
    02 Data Wrangling
    03 Answering business questions
    04 Exporting data


## 01 Import data and libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# Creating a path
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'

# Import orders.csv and products.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'))
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'))

df_prods.shape

## 02 Data Wrangling

### Dropping column

# Dropping 'eval_set' column from orders.csv
df_ords.drop(columns = ['eval_set'])

# Dropping 'eval_set' column from orders.csv irrevocable
df_ords = df_ords.drop(columns = ['eval_set'])

# Checking head afer column drop
df_ords.head

# Counting values of 'days_since_prior_order'
df_ords['days_since_prior_order'].value_counts(dropna=False)

### Renaming column

# renaming column
df_ords.rename(columns = {'order_dow':'orders_day_of_week'}, inplace = True)

df_ords.head(30)

# Checking key figures
df_ords.describe()

### Changing data type

# Changing a variable’s data type
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Checking the change of data type
df_ords['order_id'].dtype

# Checking orders key figures
df_ords.describe()

### Transposing data

# Import departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'))



df_dep.head()

#Transposing df_dep
df_dep.T

# creating a new, transposed dataframe df_dep_t
df_dep_t = df_dep.T

df_dep_t

# Adding an index to df_dep_t
df_dep_t.reset_index()

# Take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

new_header

# Take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# set the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

# Defining a data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

# create a new df snacks with department _id = 19
df_snacks = df_prods.loc[df_prods['department_id']==19]

## 03 Answering business questions

### What is the busiest hour for placing orders?

# What is the busiest hour for placing orders
df_ords['order_hour_of_day'].value_counts()

### Determine the meaning of the value 4 in 'department_id'


data_dict

print(data_dict.get('4'))

### The sales team in your client’s organization wants to know more about breakfast item sales.

# Checking prods head
df_prods.head()

# Checking ords head
df_ords.head()

# Defining a subset with only breakfast products
df_breakfast = df_prods.loc[df_prods['department_id'].isin([14])]

# Checking subsets head
df_breakfast.head()

### They’d also like to see details about products that customers might use to throw dinner parties

# Creating a subset of products for dinner parties
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

# Checking new subsets head
df_dinner_parties.head()

### How many products for dinner parties?

# Checking info of dinner_parties subset
df_dinner_parties.info()

# Checking the shape
df_dinner_parties.shape

### Extract all information about the customer with the user_id 1

# Picking that user
df_ords['user_id']='1'

# Creating a subset of data with data of user 1
df_user_1=df_ords.loc[df_ords['user_id'].isin(['1'])]

# checking head of subset
df_user_1.head()

# Checking subsets info
df_user_1.info()


# Reformating user id in int
df_user_1['user_id'] = df_user_1['user_id'].astype('int')

### You also need to provide some details about this user’s behavior

# Checking key figures of user 1
df_user_1.describe()

# Changing user_id's data type
df_ords['user_id'] = df_ords['user_id'].astype('int')

# Checking key figures of all orders
df_ords.describe()

## 04 Exporting data

df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
