==> Instacart Basket Analysis/03 Scripts/02 Consistency Checks.ipynb <==
# 02 Data Consistency Checks

    01 Cleaning and Consistency Checks for orders data
    02 Cleaning and Consistency Checks for product data 
    03 Data Export


# Import libraries
import pandas as pd
import numpy as np
import os

# Creating a path
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'

# Import orders.csv and products.csv
df_ords = pd.read_csv(os.path.join(path,'02 Data', 'Prepared Data', 'orders_wrangled.csv'))
df_prods = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'products.csv'))



## 01 Cleaning and Consistency Checks for orders data

        Consistency check
        Missing values
        Checking for mixed-type data
        Checking for duplicates

### Consistency check

# Consistency Check
df_ords.describe()


#### This all looks fine to me. 0-23 hours of the day, days since prior order is between 0 and 30 sounds reasonable, 0-6 days of a week

### Missing values

# Finding missing values
df_ords.isnull().sum()

#### The missing data makes sense, since a first-time customers cannot have reasonabele data in it

### Checking for mixed-type data

# Check for mixed types
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

### Checking for duplicates

# checking for duplicates
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups.info()

#### df_ords doesn't contain any duplicates

df_ords.shape

## 02 Cleaning and Consistency Checks for product data 

    Consistency check
    Missing value
    Subset of clean data
    Checking for duplicates in subset
    Clean subset without duplicates
    

### Consistency check

df_prods.shape

df_prods.describe()

### Missing values

# Finding missing values
df_prods.isnull().sum()

# Creating a subset of data with missing values
df_nan=df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

### Subset of clean data

# Creating a clean df
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

### Checking for duplicates in subset

#Creating a subset of duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

### Clean subset without duplicates

# Creating a clean df with no duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods.describe()

#### Checking reasonability of key figures
a price of 99.999$ seems very high. 75% goes with 11. Further investigations suggested. The count and the high std makes it somehow plausible.

## 03 Export the data

df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_final.csv'))

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'prods_final.csv'))
