==> Instacart Basket Analysis/03 Scripts/03 Combining & Exporting Data.ipynb <==
# 03 Combining & Exporting Data

    01 Import the orders_products_combined dataframe
    02 Check the shape of the imported dataframe
    03 Merging orders_products_combined dataframe with your products data set
    04 Confirm the results of the merge using the merge flag
    05 Export the newly created dataframe

    

# Import libraries
import pandas as pd
import numpy as np
import os

### 01 Import the orders_products_combined dataframe

# Creating a path
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'

# importing the data
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

### 02 Check the shape of the imported dataframe


# Checking the shape
df_ords_prods.shape

# Checking the columns
df_ords_prods.columns

df_ords_prods

# Dropping the _merge column because I had trouble with the second merged df
df_ords_prods = df_ords_prods.drop(columns = ['_merge'])

### 03 Merging orders_products_combined dataframe with your products data set


# Import products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'prods_final.csv'))

# Checking the data
df_prods.shape

# Dropping a unneccessary column
df_prods = df_prods.drop(columns = ['Unnamed: 0'])

# Merging the merged orders_products data with the data of the products 
df_orders_products_merged = df_prods.merge(df_ords_prods, on = 'product_id',how = 'inner', indicator = True,)

### 04 Confirm the results of the merge using the merge flag


df_orders_products_merged.head()

df_orders_products_merged.value_counts('_merge')

df_prods.shape

### 05 Export the newly created dataframe

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))
