==> Instacart Basket Analysis/03 Scripts/04 Deriving New Variables.ipynb <==
# 04 Deriving New Variables

    01 Create 'busiest_day' column
    02 Create 'busiest_days' column
    03 Create 'Busiest hours of the day' column
    04 Export dataframe

# import libraries
import pandas as pd
import numpy as np
import os

# Creating a path
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'

# importing data
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

## 01 Create “busiest_day” column

    Create label for 'price'
    Apply label on entire data frame
    Create 'busiest_day' list
    Create column 'busiest_day'


### Create label for 'price'

# Checking Data
df_ords_prods_merge.head()

# Checking the shape
df_ords_prods_merge.shape

# Creating a subset with the first million rows
df = df_ords_prods_merge[:1000000]

# Checking the shape of the subset
df.shape

# Define a function
def price_label(row):
    if row['prices'] <= 5:
        return 'Low range product'
    elif (row['prices']> 5) and (row['prices'] <= 15):
        return 'Mid range product'
    elif row['prices'] > 15:
        return 'High range product'
    else: return 'Not enough data'

# Applying the function
df['price_range'] = df.apply(price_label, axis=1)

# Counting the values in the column
df['price_range'].value_counts(dropna = False)

# max price
df['prices'].max()

# Labeling over again, this time with the loc function
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High range product'

# Labeling the Mid-range products
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# Labeling the low_range products
 
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

###  Apply label on entire data frame

# Labeling the High-range products in the entire dataframe
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15,'price_range_loc']= 'High-range product'

# Labeling the Mid-range products in the entire dataframe
df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5),'price_range_loc']= 'Mid-range product'

# Labeling the Low-range products in the entire dataframe
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5,'price_range_loc']= 'Low-range product'

# Counting the values of the new column
df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

### Create 'busiest_day' list

# Counting the values of each day
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

# Creating a list
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

### Create column 'busiest_day'

df_ords_prods_merge.dtypes

# Creating a new column
df_ords_prods_merge['busiest_day'] = result

# Counting the values of the column busiest_day
df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

## 02 Create'Busiest days' column
Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.


    Create list 'Busiest days'
    Create column 'busiest_days'
    Checking new column

### Create list 'Busiest days'

# Creating a new list
result1 = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result1.append("Busiest days")
  elif value == 4 or value == 3:
    result1.append("Slowest days")
  else:
    result1.append("Regularly busy")

### Create column 'busiest_days'

# Creating a new column
df_ords_prods_merge['busiest_days'] = result1

### Checking new column
Check the values of this new column for accuracy. Note any observations in markdown format.


# Conting the values in column busiest_days
df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

# Looking for the shape to compare the number of rows and the sum
df_ords_prods_merge.shape

# Comparing added number of rows to shape
result = 12916111 + 11864412 + 7624336 
result

#### Looks fine to me

## 03 Create 'Busiest hours of the day' column
When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

    Create 'busy_hour_list'
    Create column 'busiest_hours'

### Create 'busy_hour_list'

# Check value counts in "order_hour_of_day" column.

df_ords_prods_merge['order_hour_of_day'].value_counts()

# Creating a new list
busy_hour_list = []

for value in df_ords_prods_merge['order_hour_of_day']:
  if value == 10 or value == 11 or value == 14 or value == 15 or value == 13 or value == 12 or value == 16 or value == 9:
   busy_hour_list.append("Most orders")
  elif value == 3 or value == 4 or value == 2 or value == 5 or value == 1 or value == 0 or value == 6 or value == 23:
    busy_hour_list.append("Fewest orders")
  else:
    busy_hour_list.append("Average orders")

### Create column 'busiest_hours'

# Creating a new column
df_ords_prods_merge['busiest_hours'] = busy_hour_list

# Check value counts in new column.

df_ords_prods_merge['busiest_hours'].value_counts()

## 04 Export dataframe

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_derived.pkl'))
