==> Instacart Basket Analysis/03 Scripts/05 Grouping & Aggregating Variables.ipynb <==
# 05 Grouping & Aggregating Variables

    01 Import merged and derived data
    02 Aggregated mean of the “order_number” column grouped by “department_id” 
    03 Create loyalty flag for existing customers 
    04 Descriptive statistics by loyalty
    05 Create spending_flag
    06 Create order_frequency flag
    07 Export data

# import libraries
import pandas as pd
import numpy as np
import os

## 01 Import data

# Creating a path
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'

# import orders_products_merged_derived
opm=pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge_derived.pkl'))

# Creating a subset of the first million rows
df=opm[:1000000]


# Checking shape
df.shape

# Checking head
opm.head(10)

# Group by products
df.groupby('product_name')

# Grouping by products and calculating the average number of orders
df.groupby('department_id').agg({'order_number': ['mean']})

# Checking the sahpe
df.shape

# repeat the calculation, adding min and max number of orders
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

opm.head(15)

opm.head(100)

## 02 Aggregated mean of the “order_number” column grouped by “department_id” 

opm.groupby('department_id').agg({'order_number': ['mean']})

## 03 Create loyalty flag for existing customers 
using the transform() and loc() functions


#### Transform function

# creating an new column max_order and calculating the results
opm['max_order'] = opm.groupby(['user_id'])['order_number'].transform(np.max)

#### loc fuctions

# Creating a red flag for customers with > 40 orders
opm.loc[opm['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# Creating a red flag for customers with at least 11 orders and max 40
opm.loc[(opm['max_order'] <= 40) & (opm['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# Creating a red flag for customers with 10 or less ordered products
opm.loc[opm['max_order'] <=10, 'loyalty_flag'] = 'New Customer'

# Checking the head for 3 columns
opm[['user_id', 'loyalty_flag', 'order_number']].head(60)

# Counting the values of loyalty_flag
opm['loyalty_flag'].value_counts(dropna = False)

## 04 Descriptive statistics by loyalty
The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.


# Checking for the right column
opm.columns

# Grouping the data within the loyalty_flags and calculate the mean
opm.groupby('loyalty_flag').agg({'prices': ['mean', 'max']})

#### In average loyal customer buy less expensive products than New and Regular customers represented by an mean of 10.39$ New Customers spend in average 13.29$ on their products, followed by regular customers with an avergae of 12.50$

## 05 Create spending_flag
The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:

    If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”
    If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”


# Creating a new column mean_price
opm['mean_price']= opm.groupby(['user_id'])['prices'].transform('mean')

opm.head()

# Creating a new column 'spendings'
opm.loc[opm['mean_price'] < 10, 'spendings'] = 'Low spender'

# Creating a new column 'spendings'
opm.loc[opm['mean_price'] >= 10, 'spendings'] = 'High spender'

# Checking 'spendings' value_counts
opm['spendings'].value_counts()

## 06 Create order_frequency flag
In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:

    If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
    If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
    If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”


# Creating a column 'frequency median'
opm['frequency_median']=opm.groupby(['user_id'])['days_since_prior_order'].transform('median')

# Checking dataframes head
opm.head()

# Creating a new column order_frequency and assigning 'Non-frequent customer'

opm.loc[opm['frequency_median'] > 20, 'order_frequency'] = 'Non-frequent customer'

# Creating a new column order_frequency and assigning 'Regular customer'
opm.loc[(opm['frequency_median'] <= 20) & (opm['frequency_median'] > 10), 'order_frequency'] = 'Regular customer'

# Creating a new column order_frequency and assigning 'Frequent customer'

opm.loc[opm['frequency_median'] <= 10, 'order_frequency'] = 'Frequent customer'

# Checking 'order_frequency's value_counts
opm['order_frequency'].value_counts()

## 07 Export data

# Export data
opm.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_with_flags.pkl'))
