==> Instacart Basket Analysis/03 Scripts/08 Minimize size of ords_prods_customers.ipynb <==
# 08 Minimizing the size of ords_prods_customers

    01 Import data
    02 Changing int64 in int 32
    03 Changing float64 in float32
    04 Checking changes
    05 Export data frame

## 01 Import data

# import libraries

import pandas as pd
import numpy as np
import os


# Import dataframe
path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\4. Python Fundamentals for Data Analysts\Instacart Basket Analysis - Kopie'
df = pd.read_pickle(os.path.join(path,'02 Data', 'Prepared Data', 'ords_prods_customers.pkl'))

df.head()

df.dtypes

## 02 Changing int64 in int 32

df['product_id']=df['product_id'].astype('int32')

df['aisle_id']=df['aisle_id'].astype('int32')

df['department_id']=df['department_id'].astype('int32')

df['age']=df['age'].astype('int32')

df['n_dependants']=df['n_dependants'].astype('int32')

df['income']=df['income'].astype('int32')

## 03 Changing float64 in float32

df['order_number']=df['order_number'].astype('float32')

df['prices']=df['prices'].astype('float32')

df['order_id']=df['order_id'].astype('float32')

df['user_id']=df['user_id'].astype('float32')

df['orders_day_of_week']=df['orders_day_of_week'].astype('float32')

df['order_hour_of_day']=df['order_hour_of_day'].astype('float32')

df['days_since_prior_order']=df['days_since_prior_order'].astype('float32')

df['add_to_cart_order']=df['add_to_cart_order'].astype('float32')

df['reordered']=df['reordered'].astype('float32')

df['max_order']=df['max_order'].astype('float32')

df['mean_price']=df['mean_price'].astype('float32')

df['frequency_median']=df['frequency_median'].astype('float32')

## 04 Checking changes

## Checking changed data types
df.info()

## 05 Export data frame

# save the merged dataframe as a pickle
df.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_customers_reduced_size.pkl'))
